Consolidate handling cosmo_arrays in visualisation and big namespace cleanup.
