+++ swiftsimio/visualisation/__init__.py
@@ -2,7 +2,8 @@
 Visualisation sub-module for swiftismio.
 """
 
-from .projection import scatter, project_gas, project_gas_pixel_grid
-from .slice import slice_scatter as slice
-from .slice import slice_gas, slice_gas_pixel_grid
+from .projection import project_gas
+from .slice import slice_gas
+from .volume_render import render_gas
+from .ray_trace import panel_gas
 from .smoothing_length import generate_smoothing_lengths
