+++ swiftsimio/visualisation/slice_backends/__init__.py
@@ -2,6 +2,8 @@
 Backends for density slicing.
 """
 
+from swiftsimio.visualisation._vistools import backends_restore_cosmo_and_units
+
 from swiftsimio.visualisation.slice_backends.sph import (
     slice_scatter as sph,
     slice_scatter_parallel as sph_parallel,
@@ -11,9 +13,12 @@ from swiftsimio.visualisation.slice_backends.nearest_neighbours import (
     slice_scatter_parallel as nearest_neighbours_parallel,
 )
 
-backends = {"sph": sph, "nearest_neighbours": nearest_neighbours}
+backends = {
+    "sph": backends_restore_cosmo_and_units(sph),
+    "nearest_neighbours": backends_restore_cosmo_and_units(nearest_neighbours),
+}
 
 backends_parallel = {
-    "sph": sph_parallel,
-    "nearest_neighbours": nearest_neighbours_parallel,
+    "sph": backends_restore_cosmo_and_units(sph_parallel),
+    "nearest_neighbours": backends_restore_cosmo_and_units(nearest_neighbours_parallel),
 }
